logging/grafana-prometheus-loki/requirements.txt
fastapi
uvicorn

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

prometheus_client